Fix open picker on invalid types
