+++ FLUTTER_VERSION
@@ -1,1 +1,1 @@
-3.16.4
+3.16.5
