+++ api/pubspec.yaml
@@ -12,7 +12,7 @@ dependencies:
   json_annotation: ^4.8.1
   replay_bloc: ^0.2.5
   rxdart: ^0.27.7
-  xml: ^6.4.2
+  xml: ^6.5.0
 dev_dependencies:
   test: ^1.25.0
   json_serializable: ^6.7.1
