--- app/lib/api/open_release_notes.dart
@@ -1,10 +1,0 @@
-import 'package:url_launcher/url_launcher.dart';
-
-Future<bool> openReleaseNotes() {
-  return launchUrl(
-      Uri(
-          scheme: 'https',
-          host: 'go.linwood.dev',
-          pathSegments: ['butterfly', '2.0']),
-      mode: LaunchMode.externalApplication);
-}
