+++ app/lib/settings/general.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:butterfly/api/open.dart';
 import 'package:butterfly/api/save.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -9,7 +10,6 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:http/http.dart' as http;
 
-import '../api/open_release_notes.dart';
 import '../widgets/window.dart';
 
 @immutable
