+++ app/lib/views/home.dart
@@ -11,7 +11,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 
-import '../api/open_release_notes.dart';
 import '../main.dart';
 import 'files/view.dart';
