+++ app/pubspec.yaml
@@ -88,7 +88,7 @@ dependencies:
       ref: 75b5ec0e51096ae169df3cd4f02c9f6affe22a81
       path: packages/networker/networker_socket
   flutter_localized_locales: ^2.0.5
-  dynamic_color: ^1.6.8
+  dynamic_color: ^1.6.9
   popover: ^0.2.9
   markdown: ^7.1.1
   device_info_plus: ^9.1.1
