+++ docs/package.json
@@ -24,11 +24,11 @@
     "@docusaurus/preset-classic": "3.0.1",
     "@docusaurus/theme-common": "3.0.1",
     "@mdx-js/react": "^3.0.0",
-    "@swc/core": "^1.3.100",
+    "@swc/core": "^1.3.101",
     "animate.css": "^4.1.1",
     "clsx": "^2.0.0",
     "node-fetch": "^3.3.2",
-    "prism-react-renderer": "^2.3.0",
+    "prism-react-renderer": "^2.3.1",
     "react": "^18.2.0",
     "react-animation-on-scroll": "^5.1.0",
     "react-dom": "^18.2.0",
