+++ fastlane/metadata/android/en-US/changelogs/84.txt
@@ -2,6 +2,7 @@
 * Change default scale of label tool to 2
 * Improve responsiveness of dialogs on mobile devices ([#550](https://github.com/LinwoodDev/Butterfly/issues/550))
 * Improve add dialog trailing icon ([#550](https://github.com/LinwoodDev/Butterfly/issues/550))
+* Replace file picker package
 * Fix back button closes app on mobile ([#551](https://github.com/LinwoodDev/Butterfly/issues/551))
 
 View all changes in the blog: https://linwood.dev/butterfly/2.0.0-rc.3
